web: extract ProgressBackdrop into separate module
Moves ProgressBackdrop and its test out of Page to reduce complexity and
allow reuse by other components.
